requirements.txt
streamlit
torch
torchvision
timm
numpy
pillow
matplotlib
gdown
